refactor: cqrs구조로 분리를 위한 패키지 변경 및 서비스명 변경
